refactor: 💡 move workers/splits to workers/datasets_based
to prepare a generic worker
