+++ .vscode/monorepo.code-workspace
@@ -33,12 +33,8 @@
       "path": "../workers/first_rows"
     },
     {
-      "name": "workers/parquet",
-      "path": "../workers/parquet"
-    },
-    {
-      "name": "workers/splits",
-      "path": "../workers/splits"
+      "name": "workers/datasets_based",
+      "path": "../workers/datasets_based"
     }
   ],
   "settings": {
