+++ workers/datasets_based/Makefile
@@ -1,5 +1,5 @@
 # environment variables for the commands (docker compose, poetry)
-export COMPOSE_PROJECT_NAME := splits
+export COMPOSE_PROJECT_NAME := datasets_based
 export MONGO_PORT := 27040
 export CACHE_MONGO_URL := mongodb://localhost:${MONGO_PORT}
 export QUEUE_MONGO_URL := mongodb://localhost:${MONGO_PORT}
@@ -12,9 +12,9 @@ include ../../tools/Docker.mk
 
 .PHONY: run
 run:
-	poetry run python src/splits/main.py
+	poetry run python src/datasets_based/main.py
 
 .PHONY: pip-audit
 pip-audit:
-	bash -c "poetry run pip-audit -r <(poetry export -f requirements.txt --with dev | sed '/^pymongo==/,+110 d' | sed '/^requests==2.28.1 ;/,+2 d' | sed '/^kenlm @/d' | sed '/^trec-car-tools @/d')"
+	bash -c "poetry run pip-audit -r <(poetry export -f requirements.txt --with dev | sed '/^pymongo==/,+110 d' | sed '/^requests==2.28.1 ;/,+2 d' | sed '/^kenlm @/d' | sed '/^trec-car-tools @/d' | sed '/^libcommon @/,+1 d')"
 # ^ we remove problematic lines to have a working pip-audit. See https://github.com/pypa/pip-audit/issues/84#issuecomment-1326203111 for "requests"
