+++ workers/datasets_based/tests/fixtures/data/test_image_rgb.jpg
